% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{backtracking_line_search}
\alias{backtracking_line_search}
\title{Backtracking line search for gradient decent}
\usage{
backtracking_line_search(
  x,
  grad_x,
  f,
  alpha = 0.25,
  beta = 0.5,
  max_iter = 1000L,
  ...
)
}
\arguments{
\item{x}{numeric vector for x in dom f}

\item{grad_x}{numeric vector for gradient f}

\item{f}{R function for the objective function}

\item{alpha}{numeric scalar between (0, 0.5)}

\item{beta}{numeric scalar between (0, 1)}
}
\value{
numeric value for the step_size
}
\description{
Backtracking line search for gradient decent
}
\examples{
set.seed(2023)
regdat <- simul_linear_reg(regcoef = c(rep(c(1,-1),length=10),rep(0,length=90)),
intercept = 2, 
n = 50, R_sq = 0.9)
lambda <- 0.04
n = length(regdat$y)
p = ncol(regdat$X)
b <- rnorm(p+1)
X1 <- cbind(1,regdat$X)
grad_b <- crossprod(X1, X1\%*\%b - regdat$y)/n + lambda*b
f <- function(x) return(ridge_reg_loss(x[1],x[-1],lambda,regdat$y,regdat$X))
step_size <-  backtracking_line_search(b,grad_b,f=f)
print(c(f(b),f(b-step_size*grad_b)))
}
\author{
Jian Kang <jiankang@umich.edu>
}
