% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backtracking_line_search_cpp}
\alias{backtracking_line_search_cpp}
\title{Backtraking line search in Rcpp}
\usage{
backtracking_line_search_cpp(
  paras,
  grad_paras,
  obj_fun,
  dat,
  a = 0.25,
  b = 0.5,
  max_iter = 1000L
)
}
\arguments{
\item{paras}{current value of the parameter}

\item{grad_paras}{current value of the gradient parameter}

\item{obj_fun}{objective function to be minimized}

\item{dat}{list object for additional information for the objective function}

\item{a}{numeric scalar between (0, 0.5)}

\item{b}{numeric scalar between (0, 1)}
}
\value{
numeric value for the step_size
}
\description{
Backtraking line search in Rcpp
}
\author{
Jian Kang <jiankang@umich.edu>
}
