% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{Barzilai_Borwein_method}
\alias{Barzilai_Borwein_method}
\title{Barzilai Borwein method to determine the step size for gradient descent}
\usage{
Barzilai_Borwein_method(x, prev_x, grad_x, prev_grad_x, step_size = 1e-06)
}
\arguments{
\item{x}{numeric vector for current parameters}

\item{prev_x}{numeric vector for previous parameters}

\item{grad_x}{numeric vector for current gradient}

\item{prev_grad_x}{vector for previous gradient}

\item{step_size}{lower bound of the step size}
}
\value{
numeric scalar for step size
}
\description{
Barzilai Borwein method to determine the step size for gradient descent
}
\examples{
set.seed(2023)
regdat <- simul_linear_reg(regcoef = c(rep(c(1,-1),length=5),rep(0,length=5)),
intercept = 2, 
n = 500, R_sq = 0.9)
lambda <- 0.04
step_size <- 0.001
n = length(regdat$y)
p = ncol(regdat$X)
b <- rnorm(p+1)
X1 <- cbind(1,regdat$X)
resid <- regdat$y - X1\%*\%b
grad_b <- crossprod(X1,-resid)/n + lambda*b
prev_b <- b
b <- b - step_size*grad_b
resid <- regdat$y - X1\%*\%b
prev_grad_b <- grad_b
grad_b <- crossprod(X1,-resid)/n + lambda*b
f <- function(x) return(ridge_reg_loss(x[1],x[-1],lambda,regdat$y,regdat$X))
prev_step_size <- step_size
step_size <-  Barzilai_Borwein_method(b, prev_b, grad_b, prev_grad_b)
print(c(f(b+prev_step_size*prev_grad_b),f(b),f(b-step_size*grad_b)))
}
\author{
Jian Kang <jiankang@umich.edu>
}
