% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_ridge_reg_loss_cpp}
\alias{grad_ridge_reg_loss_cpp}
\title{compute the gradient of the ridge regression loss in Rcpp}
\usage{
grad_ridge_reg_loss_cpp(beta, dat, idx_R = as.integer(c()))
}
\arguments{
\item{beta}{regression coefficients}

\item{dat}{list object containing covariate matrix, response variable}

\item{idx_R}{vector of positive integers for sample indices.}
}
\value{
a vector of gradient of the loss function
}
\description{
compute the gradient of the ridge regression loss in Rcpp
}
\author{
Jian Kang <jiankang@umich.edu>
}
