% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradient_descent_cpp}
\alias{gradient_descent_cpp}
\title{Gradient descent algorithm with different step size methods}
\usage{
gradient_descent_cpp(
  paras,
  obj_fun,
  grad_fun,
  dat,
  step_size_method = as.character(c("fixed", "back", "BB")),
  max_iter = 100000L,
  step_size = 1e-04,
  tol = 1e-04,
  verbose = 100L
)
}
\arguments{
\item{paras}{numeric vector}

\item{obj_fun}{objective function to be minimized}

\item{grad_fun}{gradient function}

\item{dat}{list object to pass other data to the objective function 
and the gradient of the objective functions}

\item{step_size_method}{indicate the methods to determine the step size}

\item{max_iter}{positive integer for maximum number of iterations}

\item{step_size}{numeric scalar for the step size}

\item{tol}{tolerance for the stopping criterion}
}
\value{
a list of objects
}
\description{
Gradient descent algorithm with different step size methods
}
\examples{
regdat <- simul_linear_reg(n = 1000, p = 5, pos_corr = 0.6)
regdat$lambda = 0.001
res <- gradient_descent_cpp(paras=rep(0,length=ncol(regdat$X)),
ridge_reg_loss_cpp,grad_ridge_reg_loss_cpp,regdat)
}
\author{
Jian Kang <jiankang@umich.edu>
}
