% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{linear_ridge_reg}
\alias{linear_ridge_reg}
\title{Fit ridge regression by different methods}
\usage{
linear_ridge_reg(
  y,
  X,
  intercept = TRUE,
  lambda = 0.001,
  method = c("solve", "solve-large-p", "gd", "gd-exact", "gd-back", "gd-BB", "sgd",
    "sgd-decrease", "sgd-batch"),
  initial_regcoef = rep(0, length = ncol(X)),
  initial_intercept = 0,
  step_size = 1e-04,
  max_iter = 100000L,
  tol = 1e-04,
  alpha = 0.25,
  beta = 0.5,
  diminishing_ratio = 0.1,
  r = 0.55,
  batch_size = 10,
  verbose = Inf,
  comp_loss = FALSE
)
}
\arguments{
\item{y}{numeric vector for n observations of the response variable}

\item{X}{covariate matrix of dimension n by p}

\item{intercept}{logical variable indicating whether include the intercept 
in the model. The default value is TRUE}

\item{lambda}{tuning parameter}

\item{method}{character indicating one of the following methods 
c("solve","solve-large-p","gd","gd-exact","gd-back","gd-BB"). 
"solve": solving linear system of dimensions p by p
"solve-large-p": solving linear system of dimensions n by n
"gd": Gradient descent with a fixed step size
"gd-exact": Gradient descent with exact line search 
"gd-back": Gradient descent with backtracking line search
"gd-BB": Gradient descent with Barzilai Borwein method to update the step size
"sgd": Stochastic gradient descent with a fixed step size}
}
\value{
list objects with the following elements
\describe{
\item{intercept}{intercept estimate}
\item{regcoef}{a vector of p regression coefficient estimates}
\item{residuals}{a vector of n residuals}
\item{fitted.values}{the fitted mean values}
\item{gd}{list object including convergence information}
\item{elapsed}{computing time in seconds}
\item{loss}{value of loss function at the parameter estimates}
}
}
\description{
Fit ridge regression by different methods
}
\examples{
library(glmnet)
set.seed(2023)
betacoef <- c(rep(c(1,-1),length=10),rep(0,length=200 - 10))
dat <- simul_linear_reg(regcoef = betacoef,intercept=2,
n = 10000,R_sq = 0.9)
lambda <- 0.05
method_list <- c("solve","solve-large-p","gd","gd-exact","gd-back","gd-BB","sgd")
res_list <- list()
loss_list <- list()
mse_list <- list()
elapsed_list <- list()
for(method in method_list){
  res_list[[method]] <- with(dat, linear_ridge_reg(y = y, X = X, lambda = lambda, method=method))
  loss_list[[method]] <- res_list[[method]]$loss
  mse_list[[method]] <- mean((res_list[[method]]$regcoef - dat$regcoef)^2)
  elapsed_list[[method]] <- as.numeric(res_list[[method]]$elapsed)
}
elapsed_list[["glmnet"]] <- proc.time()[3]
res_list[["glmnet"]] <- with(dat,glmnet(x=X, y=y, alpha=0.0, lambda = lambda))
elapsed_list[["glmnet"]] <- proc.time()[3] - elapsed_list[["glmnet"]]
loss_list[["glmnet"]] <- with(res_list[["glmnet"]],ridge_reg_loss(intercept=a0,regcoef=beta,lambda=lambda,dat$y,dat$X))
mse_list[["glmnet"]] <- mean((res_list[["glmnet"]]$beta - dat$regcoef)^2)
loss_list[["true"]] <- with(dat,ridge_reg_loss(intercept,regcoef,lambda=lambda,y,X))
mse_list[["true"]] <- NA
elapsed_list[["true"]] <- NA
print(cbind(loss = unlist(loss_list),
           mse = unlist(mse_list),
           elapsed = unlist(elapsed_list)))
}
\author{
Jian Kang <jiankang@umich.edu>
}
