% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{match_paras}
\alias{match_paras}
\title{find the column index permutation to match two matrix parameters}
\usage{
match_paras(target_paras, paras)
}
\arguments{
\item{target_paras}{matrix of parameters}

\item{paras}{matrix of parameters}
}
\value{
column indices idx_match such that target_paras is matched with paras[,idx_match]
}
\description{
find the column index permutation to match two matrix parameters
}
\examples{
library(flexmix)
set.seed(2024)
dat <- simul_mixture_reg(n = 1000, K = 3)
fit <- flexmix(dat$y~0+dat$X,k = 3)
alpha_est <- parameters(fit)[1:ncol(dat$X),]
print(alpha_est) 
idx <- match_paras(dat$alpha,alpha_est)
print(dat$alpha)
print(alpha_est[,idx])
print(idx)
}
\author{
Jian Kang <jiankang@umich.edu>
}
