% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{ridge_reg_exact_line_search}
\alias{ridge_reg_exact_line_search}
\title{Exact line search for ridge regression}
\usage{
ridge_reg_exact_line_search(
  beta,
  Delta_beta,
  lambda,
  X,
  resid,
  step_size_0 = 0
)
}
\arguments{
\item{beta}{regression coefficients}

\item{Delta_beta}{descent direction for regression coefficients}

\item{lambda}{tuning parameter}

\item{X}{design matrix}

\item{resid}{residuals}
}
\value{
step_size
}
\description{
Exact line search for ridge regression
}
\examples{
set.seed(2023)
regdat <- simul_linear_reg(regcoef = c(rep(c(1,-1),length=10),rep(0,length=90)),
intercept = 2, 
n = 50, R_sq = 0.9)
lambda <- 0.04
n = length(regdat$y)
p = ncol(regdat$X)
b <- rnorm(p+1)
X1 <- cbind(1,regdat$X)
resid <- regdat$y - X1\%*\%b
grad_b <- crossprod(X1,-resid)/n + lambda*b
f <- function(x) return(ridge_reg_loss(x[1],x[-1],lambda,regdat$y,regdat$X))
step_size <-  ridge_reg_exact_line_search(b, -grad_b, lambda, X1, resid)
print(c(f(b),f(b-step_size*grad_b)))
}
\author{
Jian Kang <jiankang@umich.edu>
}
