% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ridge_reg_loss_cpp}
\alias{ridge_reg_loss_cpp}
\title{Compute loss of the ridge regression in Rcpp}
\usage{
ridge_reg_loss_cpp(beta, dat, idx_R = as.integer(c()))
}
\arguments{
\item{beta}{numeric vector for regression coefficients}

\item{dat}{list object containing covariate matrix, response variable}

\item{idx_R}{vector of positive integers for sample indices.}
}
\value{
value of the loss function
}
\description{
Compute loss of the ridge regression in Rcpp
}
\author{
Jian Kang <jiankang@umich.edu>
}
