% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{simul_covar_mat}
\alias{simul_covar_mat}
\title{simulate design matrix with a positive 
exchangeable correlation from 
a multivariate normal distribution}
\usage{
simul_covar_mat(n, p, pos_corr = 0.5, sd = 1)
}
\arguments{
\item{n}{positive integer for sample size}

\item{p}{positive integer for number of predictors}

\item{pos_corr}{positive value in (0,1) for positive correlation coefficient}

\item{sd}{positive numeric scalar for the standard deviation}
}
\value{
a covariate matrix of dimension n by p
}
\description{
simulate design matrix with a positive 
exchangeable correlation from 
a multivariate normal distribution
}
\examples{
X <- simul_covar_mat(n = 1000, p = 5, pos_corr = 0.5, sd = 2)
print(cor(X))
print(cov(X))
}
\author{
Jian Kang <jiankang@umich.edu>
}
