% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{simul_linear_reg}
\alias{simul_linear_reg}
\title{simulate data for a linear regression}
\usage{
simul_linear_reg(
  regcoef = NULL,
  X = NULL,
  intercept = 0,
  sd_eps = NULL,
  R_sq = 0.9,
  n = 1000,
  p = length(regcoef),
  pos_corr = 0.5,
  sd_X = 1
)
}
\arguments{
\item{regcoef}{numeric vector for regression coefficients}

\item{X}{numeric matrix of dimension}

\item{sd_eps}{positive numeric scalar for the random noise standard deviation}

\item{R_sq}{positive value in (0,1) for coefficient of 
determination (R-squared) if sd_eps = NULL}

\item{n}{positive integer for sample size if X is NULL}

\item{p}{positive integer for number of covariates if X is NULL}

\item{pos_corr}{positive value in (0,1) for positive correlation between 
covariates if X is NULL}

\item{sd_X}{positive numeric scalar for the standard deviation of 
covariates if X is NULL}
}
\value{
list object of 
\describe{
\item{y}{vector of n response variables}
\item{X}{covariate matrix}
\item{regcoef}{vector of p regression coefficients}
\item{intercept}{numeric scalar}
\item{mu}{vector of E(y | X)}
\item{sd_eps}{noise standard deviation}
\item{R_sq}{R squared}
}
}
\description{
simulate data for a linear regression
}
\examples{
regdat <- simul_linear_reg(n = 1000, p = 5, pos_corr = 0.6)
res <- with(regdat, lm(y~X))
print(cbind(true_coef = c(regdat$intercept,regdat$regcoef), 
           est_coef = coef(res)))
}
\author{
Jian Kang <jiankang@umich.edu>
}
