% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmichBiostat815.R
\name{simul_mixture_reg}
\alias{simul_mixture_reg}
\title{simulate Gaussian mixture regression model}
\usage{
simul_mixture_reg(
  n = 1000,
  p = 5,
  K = 3,
  pos_corr = 0.5,
  min_effect = 1,
  max_effect = 3,
  prec_prob = 50,
  R_sq = 0.9
)
}
\arguments{
\item{n}{positive integer for sample size}

\item{p}{positive integer for number of covariates}

\item{K}{positive integer for number of components}

\item{pos_corr}{numeric scalar in (0,1) for exchangable correlation between covariates}

\item{min_effect}{positive numeric scalar for the miminum effect size of the regression coefficients}

\item{max_effect}{positive numeric scalar for the maximum effect size of the regression coefficients}

\item{prec_prob}{positive numeric scalar control the weight probabilty parameters}

\item{R_sq}{numeric scalar in (0,1) for R squared}
}
\value{
list object of 
\describe{
\item{y}{vector of n response variables}
\item{X}{covariate matrix}
\item{alpha}{p by K matrix for regression coefficients}
\item{prob}{K by 1 vector for weight probabilit}
\item{sigma_sq}{positive numeric scalar for the noise variance}
\item{delta}{n by 1 vector for group labels}
\item{R_sq}{R squared}
}
}
\description{
simulate Gaussian mixture regression model
}
\examples{
library(flexmix)
dat <- simul_mixture_reg(n = 1000, K = 3)
fit <- flexmix(dat$y~0+dat$X,k = 3)
paras <- parameters(fit)[1:ncol(dat$X),]
print(paras) 
}
\author{
Jian Kang <jiankang@umich.edu>
}
